+++ player/tests/test.rs
@@ -3,7 +3,7 @@
 //!  and run the tests through them.
 //!
 //!  Test requirements:
-//!    - all IDs have the backend `Empty`
+//!    - all IDs have the backend `Noop`
 //!    - all expected buffers have `MAP_READ` usage
 //!    - last action is `Submit`
 //!    - no swapchain use
@@ -74,9 +74,7 @@ impl Test<'_> {
             wgt::Backend::Gl => "Gl",
             _ => unreachable!(),
         };
-        let string = read_to_string(&path)
-            .unwrap()
-            .replace("Empty", backend_name);
+        let string = read_to_string(&path).unwrap().replace("Noop", backend_name);
         ron::de::from_str(&string)
             .unwrap_or_else(|e| panic!("{path:?}:{} {}", e.position.line, e.code))
     }
