+++ wgpu-hal/examples/ray-traced-triangle/main.rs
@@ -1129,7 +1129,7 @@ cfg_if::cfg_if! {
     }
     // Fallback
     else {
-        type Api = hal::api::Empty;
+        type Api = hal::api::Noop;
     }
 }
